Add integration test for delete

+++ tests/lambda_function/test_integration.py
@@ -25,7 +25,8 @@ def nginx_deployment_info():
     yield deployment_dict, namespace, name
 
     apps_v1_api = kubernetes.client.AppsV1Api()
-    apps_v1_api.delete_namespaced_deployment(namespace=namespace, name=name)
+    if apps_v1_api.list_namespaced_deployment(namespace=namespace).items:
+        apps_v1_api.delete_namespaced_deployment(namespace=namespace, name=name)
 
 
 @pytest.mark.integration
@@ -83,6 +84,27 @@ def test_deployment_update(nginx_deployment_info, _nginx_deployment):
     assert response.items[0].metadata.labels["key"] == "value"
 
 
+@pytest.mark.integration
+def test_deployment_delete(nginx_deployment_info, _nginx_deployment):
+    """
+    GIVEN deployment as dictionary that has been created
+    WHEN delete is called with the deployment dictionary as a body and physical name
+    THEN the deployment is deleted.
+    """
+    deployment_dict, namespace, name = nginx_deployment_info
+
+    result = operations.delete(
+        body=deployment_dict, physical_name=f"{namespace}/{name}"
+    )
+
+    # Check result
+    assert result.status == "SUCCESS"
+    # Check that the deployment is deleted
+    apps_v1_api = kubernetes.client.AppsV1Api()
+    response = apps_v1_api.list_namespaced_deployment(namespace=namespace)
+    assert not response.items
+
+
 @pytest.fixture
 def cluster_role_info():
     """Kubernetes cluster role."""
@@ -94,7 +116,10 @@ def cluster_role_info():
     yield cluster_role_dict, name
 
     rbac_v1_api = kubernetes.client.RbacAuthorizationV1Api()
-    rbac_v1_api.delete_cluster_role(name=name)
+    response = rbac_v1_api.list_cluster_role()
+    name_mapping = {item.metadata.name for item in response.items}
+    if name in name_mapping:
+        rbac_v1_api.delete_cluster_role(name=name)
 
 
 @pytest.mark.integration
@@ -149,3 +174,23 @@ def test_cluster_role_update(cluster_role_info, _cluster_role):
     metadata_dict = {item.metadata.name: item.metadata for item in response.items}
     assert name in metadata_dict
     assert metadata_dict[name].labels["key"] == "value"
+
+
+@pytest.mark.integration
+def test_cluster_role_delete(cluster_role_info, _cluster_role):
+    """
+    GIVEN cluster role as dictionary that has been created
+    WHEN delete is called with the cluster role dictionary as a body and physical name
+    THEN the cluster role is deleted.
+    """
+    cluster_role_dict, name = cluster_role_info
+
+    result = operations.delete(body=cluster_role_dict, physical_name=name)
+
+    # Check result
+    assert result.status == "SUCCESS"
+    # Check that the cluster_role is created
+    rbac_v1_api = kubernetes.client.RbacAuthorizationV1Api()
+    response = rbac_v1_api.list_cluster_role()
+    assert response.items
+    assert name not in {item.metadata.name for item in response.items}
